backend/requirements.txt
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
numpy==1.21.4
oauthlib==3.2.0
pandas==1.3.5
psycopg2-binary==2.9.3
pylint==2.12.2
PyYAML==5.3.1
regex==2020.11.13
sqlalchemy==1.4.36
gunicorn==20.0.4
jupyter-server==1.11.2
Flask[async]==2.2.2
Flask-Cors==3.0.10
pyparsing==3.0.9
Flask-Caching==2.0.2
psutil==5.9.4
polars==0.16.16
pyarrow==11.0.0
python-box==7.0.1
fbprophet==0.7.1
yfinance==0.2.14